applied a part of the large patch from Steve Buchi
Bug List:

- Tabbing between text boxes within the same movie clip.
- Fixed a bug with finding certain fonts in Windows.
- Fixed a bug where narrow text and a small font size would not be displayed correctly.
- Fixed a bug which occassionally caused a crach when flash variables are passed in during movie loading
- Added ability to hide text in a password text box.
- Fixed shift not being recognized in text box entry
- Fixed a bug in goto_frame: sometimes when rewinding frames, execute_state_reserve would remove the incorrect object.
- Now store and pass blend mode information to the renderer

+++ tu-testbed/gameswf/gameswf_types.h
@@ -62,9 +62,9 @@ namespace gameswf
 		void	set(double color)
 		{
 			int rgb = int(color);
-			m_r = (rgb >> 16) & 0xFF;
-			m_g = (rgb >> 8) & 0xFF;
-			m_b = rgb & 0xFF;
+			m_r = (Uint8)(rgb >> 16) & 0xFF;
+			m_g = (Uint8)(rgb >> 8) & 0xFF;
+			m_b = (Uint8)rgb & 0xFF;
 			m_a = 255;
 		}
